+++ src/cv2ext/__init__.py
@@ -235,7 +235,7 @@ __all__ = [
     "tracking",
     "video",
 ]
-__version__ = "0.0.24"
+__version__ = "0.0.25"
 
 _log.info(f"Initialized cv2ext with version {__version__}")