+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+## 0.0.25 (11-07-2024)
+
+### Added
+
+- bboxes.filter_bboxes_by_region
+    Get the bounding boxes from a list of bounding boxes
+    which are contained inside the region
+
+### Changed
+
+- bboxes.nms now has conf and classid switched
+    in the bounding box entries.
+
 ## 0.0.24 (11-02-2024)
 
 ### Added